Initialchanges to add MSVCsupport, changed submodules to HTTPS
