+++ .gitmodules
@@ -1,7 +1,7 @@
 [submodule "src/utils/third_party/json"]
 	path = src/utils/third_party/json
-	url = git@github.com:nlohmann/json.git
+	url = https://github.com/nlohmann/json.git
 	branch = master
 [submodule "src/utils/third_party/cpp-base64"]
 	path = src/utils/third_party/cpp-base64
-	url = git@github.com:ReneNyffenegger/cpp-base64.git
+	url = https://github.com/ReneNyffenegger/cpp-base64.git
